DROID-3927 refactoring
